requirements.txt
requests>=2.28.0
ourgroceries>=1.5.4
pyyaml>=6.0
aiohttp>=3.8.0
